requirements.txt
qrcode[pil]
Pillow>=10.3.0
